+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Allow EntityType-specific custom toString methods for Entity.
 * Allow EntityType-specific custom display methods for Entity, to get a string for displaying in a GUI.
 * Fixed Auth method not initialised correctly. The AuthMethod can now be set on the service before calling init.
+* Fixed relations in Projects data model.
 
 
 ## Release Version 2.18